LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-pseudo/marker-intrinsic-contribution-001-expected.txt


PASS Intrinsic contribution of outside symbol marker
PASS Intrinsic contribution of outside decimal marker
PASS Intrinsic contribution of outside string marker
PASS Intrinsic contribution of outside content marker
PASS Intrinsic contribution of inside symbol marker
FAIL Intrinsic contribution of inside decimal marker assert_equals: clientWidth expected 30 but got 11
FAIL Intrinsic contribution of inside string marker assert_equals: clientWidth expected 60 but got 11
FAIL Intrinsic contribution of inside content marker assert_equals: clientWidth expected 70 but got 11
